% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.summary.R
\name{mean.summary}
\alias{mean.summary}
\title{to generate mean summary data frame}
\usage{
\method{mean}{summary}(data, by, var)
}
\arguments{
\item{data}{input dataframe}

\item{by}{input by group or groups}

\item{var}{input column which we need to summarize}
}
\value{
a dataframe
}
\description{
to generate mean summary data frame
}
\examples{
mean.summary(mtcars, rlang::exprs(cyl, vs, am), mpg)

}
